Fix subscriptions role list and enable candidate user edits
